Case:20949 Added tests for TopologicalSort and refactored
One test highlighted undesirable behaviour: a KeyNotFound error when
an item not in the enumerable is referenced in edges. The error message
was not appropriate and appeared accidental, so altered function to
gracefully ignore such references.

Whilst making this change, renamed a variable more descriptively and
added more detail to comments.