Case:16223 More cleanup of ToString() to make more robust and extensible.  You can now overload the Message property and it will be respected by the string formatter which will be useful for Translations.
The culture information has now gone as this is only relevant in translations
and will be added there.

The Log deserializer constructor is in place and to support it the only required
context value is GUID - everything else is entirely optional.  A GUID is a
minimum requirement as it also holds the timestamp and gaurantees uniqueness,
without a GUID at the very least you have no gaurantee you actually have a log,
so it seems a fair restriction.