Case:32794 Refactored FormatBuilder to include more overloads that don't attempt to parse the input, which should give significant improvements when adding non-string elements, etc.  You now have to use AppendFormat() or AppendFormatLine() when using tags.
Added FormattedStreamWriter, which can be used like StringWriter as an alternative to FormatBuilder (uses FormatBuilder under hood).  This doesn't replace FormatWriter, which allows formatting to be performed when writing to an existing TextWriter.

FormatWriter now only accepts a FormatBuilder, as this is it's sole use.  However it now offers a Func<FormatBuilder, FormatBuilder> overload which allow a slightly more fluent syntax such as:
writer.Write(b => b.AppendFormat("{ConsoleColor:Green}{layout:w80}Some text").AppendLine(" added to a builder"));

FormatBuilder now implements Clear() and IsEmpty(), making it more similar to StringBuilder, however it doesn't support other forms of manipulation (such as Remove and Insert) due to the complex nature of holding chunks rather than an actually built string.  Although this is something that could be easily added, it seems like an edge case, and a similar effect can be achieved by using the Append overloads that take an enumeration of chunks or a FormatBuilder (i.e. iterate over existing chunks skipping the ones that need to be removed).