Case:32794 Added LineWrapMode enum and added it to Layout. LayoutBuilder uses the value in the layout when terminating a line. ConsoleBuilder.UpdateLayout always sets this to PadToWrap
Fixed a bug with the first line after a SetLayout not having the layout applied, as the control is added to that first line, after the line has already been created with the previous layout. Not sure if its the best fix.

Added nice, human readable string to Layout.ToString.