Case:32794 Updated nugets. Added methods to append a Log to a LayoutBuilder. The ToString methods use the new AppendTo methods. ConsoleLogger has been converted to use a ConsoleBuilder. The other loggers still need to be changed to use Builders.
There are still a few minor issues with the format of the text produced in the AppendFormetted method, but its mostly as it was.
The context and inner exception data is currently added to a new Builder then ToStringed. That should be changed so that the chunks of the builder are added to the main builder.