#2, #3 WIP Major build system rewrite
Renamed a number of sub-folders to shorten paths.
Removed projects that form part of the build system from the CoreLibraries solution, they will be instead built automatically if not already built.  This speeds up the core solution build after the first invocation.
Major rewrite of ILMerge targets to support building of multiple merge assemblies, in preparation for move the MSBuild invocation.
Fixed a number of bugs in ILMerge and Initializer targets.

Still TODO - On invoking the build of required projects, the resulting use of the buildtasks needs to be run in a seperate MSBuild process, as this will *hopefully* reduce the locking problems previously encountered.