#3 ILMerge and ModuleInitializer now form part of Utilities NuGet
By including ILMerge and ModuleInitializer in the core Utilities NuGet it is now significantly easier to perform ModuleInitialization and/or ILMerging.

Defining the "ModuleInitializer" constant on the build will invoke the injection routine post build.

Defining a MergeAsm ItemGroup in the project (see ILMerge.targets for details) will configure post build merging.

ILMerge.exe is included in the NuGet by default, however if the ILMerge package is found in the solution directory then it will use the ILMerge found therein in preference.  This allows ILMerge to work without requiring the package to be loaded (with the associated dependency issues).