#6 Redesigned API
The CryptographyProvider now exposes CryptoStreams and ICryptoTransform objects, allowing streaming of encryption and decryption.

Because of the nature of these built-in APIs there's no realistic way of encoding length when the algorithm doesn't already do it, however, the utilities encryption/decryption methods take note of the new PreservesLength flag and will automatically append a termination byte (0xFF) when encrypting, and strip trailing 00's (and the last 0xFF) when decrypting.