#7, #9 Improved configuration path caching.  Added new XObjectExtensions.DeepEquals method.
The new XObjectExtensions.DeepEquals allows for more customisable comparisons between two XObjects, and is far more powerful than the builtin XNode.DeepEquals, which it emulates with XObjectComparisonOptions.Legacy.
As it uses stack based recursion it may well be faster for a number of the common modes.

Split the TestConfiguration into it's own test.config file to test the newly enhanced file change detection logic.