#10 WIP: String diff now uses new StringMap which is optimized for performance and memory utilisation.
The new String.ToMapped() extension method just create a StringMap using it's constructor for better encapsulation.

Not all TextModes are working yet, but reverse mapping (taking the diff result of the mapped string and mapping it back onto the original string) is now complete, so string diffing with optional comparitors is working.