#10 String diff now works for all valid TextOptions, and comparitors.
Greatly improved StringMap so that it is MUCH faster during diff and indexing.  There was no point storing the original as character array, when we could store the mapped as a character array instead, and just have the mappings in place, indexing is therefore as fast as indexing a string.

Split different text modes into different loops, although this means a litle code duplication it's much easier to understand and is slightly faster.