#7 Fixed issue with Default Collections on Configurations
OnDeserializeUnrecognizedElement was catching all elements, including the one's used to initialize collections, due to the way the underlying BCL code works, it now handles de-serialization of collections itself so that it can prevent errors for unknown nodes, and still capture them as XML elements.  This allow mixing of collection elements with additional elements.