#11 WIP, added new Text namespace and TextExtensions, which include new extension methods on character.
This is the first step of potentially implementing Unicode Annexes 14 and 29 in that it allows the rapid retrieval of the property of a character in relation to the Unicode Character Database property definitions of the text segmentation algorithms and line break algorithm.

The code downloads the latest property files from unicode.org automatically and converts the data into enumerations and a single int array that allows extremely optimised binary search.